django_template_project/accounts/serializers.py
from rest_framework import serializers

from django_template_project.accounts.models import CustomUser


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = (
            "id",
            "username",
        )
