requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
amqp==5.0.5
    # via
    #   -r requirements/base.txt
    #   kombu
ansible-base==2.10.6
    # via ansible
ansible==3.0.0
    # via -r requirements/dev.in
apipkg==1.5
    # via execnet
appdirs==1.4.4
    # via black
asgiref==3.3.1
    # via
    #   -r requirements/base.txt
    #   django
    #   djoser
attrs==20.3.0
    # via pytest
backcall==0.2.0
    # via
    #   -r requirements/base.txt
    #   ipython
billiard==3.6.3.0
    # via
    #   -r requirements/base.txt
    #   celery
black==20.8b1
    # via -r requirements/dev.in
celery==5.0.5
    # via -r requirements/base.txt
certifi==2020.12.5
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.14.5
    # via
    #   -r requirements/base.txt
    #   cryptography
chardet==4.0.0
    # via
    #   -r requirements/base.txt
    #   requests
click-didyoumean==0.0.3
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.1.6
    # via
    #   -r requirements/base.txt
    #   celery
click==7.1.2
    # via
    #   -r requirements/base.txt
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
coreapi==2.3.3
    # via
    #   -r requirements/base.txt
    #   djoser
coreschema==0.0.4
    # via
    #   -r requirements/base.txt
    #   coreapi
cryptography==3.4.6
    # via
    #   -r requirements/base.txt
    #   ansible-base
    #   social-auth-core
decorator==4.4.2
    # via
    #   -r requirements/base.txt
    #   ipython
defusedxml==0.6.0
    # via
    #   -r requirements/base.txt
    #   python3-openid
    #   social-auth-core
django-admin-display==1.3.0
    # via -r requirements/base.txt
django-dbbackup==3.3.0
    # via -r requirements/base.txt
django-extensions==3.1.1
    # via -r requirements/base.txt
django-filter==2.4.0
    # via -r requirements/base.txt
django-jazzmin==2.4.4
    # via -r requirements/base.txt
django-stubs==1.7.0
    # via -r requirements/dev.in
django-templated-mail==1.1.1
    # via
    #   -r requirements/base.txt
    #   djoser
django-webpack-loader==0.7.0
    # via -r requirements/base.txt
django==3.1.7
    # via
    #   -r requirements/base.txt
    #   django-admin-display
    #   django-dbbackup
    #   django-filter
    #   django-jazzmin
    #   django-stubs
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.6.0
    # via
    #   -r requirements/base.txt
    #   djoser
djangorestframework==3.12.2
    # via
    #   -r requirements/base.txt
    #   djangorestframework-simplejwt
djoser==2.1.0
    # via -r requirements/base.txt
drf-flex-fields==0.8.9
    # via -r requirements/base.txt
environs==9.3.1
    # via -r requirements/base.txt
execnet==1.8.0
    # via pytest-xdist
flake8-formatter-abspath==1.0.1
    # via -r requirements/dev.in
flake8==3.8.4
    # via
    #   -r requirements/dev.in
    #   flake8-formatter-abspath
idna==2.10
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==1.1.1
    # via pytest
ipython-genutils==0.2.0
    # via
    #   -r requirements/base.txt
    #   traitlets
ipython==7.20.0
    # via -r requirements/base.txt
isort==5.7.0
    # via -r requirements/dev.in
itypes==1.2.0
    # via
    #   -r requirements/base.txt
    #   coreapi
jedi==0.18.0
    # via
    #   -r requirements/base.txt
    #   ipython
jinja2==2.11.3
    # via
    #   -r requirements/base.txt
    #   ansible-base
    #   coreschema
kombu==5.0.2
    # via
    #   -r requirements/base.txt
    #   celery
markupsafe==1.1.1
    # via
    #   -r requirements/base.txt
    #   jinja2
marshmallow==3.10.0
    # via
    #   -r requirements/base.txt
    #   environs
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
mypy==0.812
    # via
    #   -r requirements/dev.in
    #   django-stubs
oauthlib==3.1.0
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
packaging==20.9
    # via
    #   ansible-base
    #   pytest
    #   pytest-sugar
parso==0.8.1
    # via
    #   -r requirements/base.txt
    #   jedi
pathspec==0.8.1
    # via black
pexpect==4.8.0
    # via
    #   -r requirements/base.txt
    #   ipython
pickleshare==0.7.5
    # via
    #   -r requirements/base.txt
    #   ipython
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.16
    # via
    #   -r requirements/base.txt
    #   click-repl
    #   ipython
psycopg2-binary==2.8.6
    # via -r requirements/base.txt
ptyprocess==0.7.0
    # via
    #   -r requirements/base.txt
    #   pexpect
py==1.10.0
    # via
    #   pytest
    #   pytest-forked
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via
    #   -r requirements/base.txt
    #   cffi
pyflakes==2.2.0
    # via flake8
pygments==2.8.0
    # via
    #   -r requirements/base.txt
    #   ipython
pyjwt==2.0.1
    # via
    #   -r requirements/base.txt
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==2.4.7
    # via packaging
pytest-django==4.1.0
    # via -r requirements/dev.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-sugar==0.9.4
    # via -r requirements/dev.in
pytest-xdist==2.2.1
    # via -r requirements/dev.in
pytest==6.2.2
    # via
    #   -r requirements/dev.in
    #   pytest-django
    #   pytest-forked
    #   pytest-sugar
    #   pytest-xdist
python-dotenv==0.15.0
    # via
    #   -r requirements/base.txt
    #   environs
python3-openid==3.2.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
pytz==2021.1
    # via
    #   -r requirements/base.txt
    #   celery
    #   django
    #   django-dbbackup
pyyaml==5.4.1
    # via ansible-base
regex==2020.11.13
    # via black
requests-oauthlib==1.3.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
requests==2.25.1
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
six==1.15.0
    # via
    #   -r requirements/base.txt
    #   click-repl
    #   django-dbbackup
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via
    #   -r requirements/base.txt
    #   djoser
social-auth-core==4.0.3
    # via
    #   -r requirements/base.txt
    #   social-auth-app-django
sqlparse==0.4.1
    # via
    #   -r requirements/base.txt
    #   django
termcolor==1.1.0
    # via pytest-sugar
toml==0.10.2
    # via
    #   black
    #   pytest
traitlets==5.0.5
    # via
    #   -r requirements/base.txt
    #   ipython
typed-ast==1.4.2
    # via
    #   black
    #   mypy
typing-extensions==3.7.4.3
    # via
    #   black
    #   django-stubs
    #   mypy
uritemplate==3.0.1
    # via
    #   -r requirements/base.txt
    #   coreapi
urllib3==1.26.3
    # via
    #   -r requirements/base.txt
    #   requests
vine==5.0.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
wcwidth==0.2.5
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
